ADD current forecast UI
Added the interface and view code for displaying the current forecast date and air quality.

+++ AirScanner/Main/Networking/Objects/ApiResponses.swift
@@ -21,7 +21,7 @@ struct APIForecastListResponse: Codable {
 struct APIForecastResponse: Codable {
     let airQualityIndex: APIAirQualityResponse
     let components: ForecastComponents
-    let date: Int
+    let date: Double
 
     enum CodingKeys: String, CodingKey {
         case airQualityIndex = "main"
